Remove ocean-legacy
