+++ squid_py/ocean.py
@@ -1,19 +1,17 @@
 import logging
-import os
 
 from web3 import Web3, HTTPProvider
 
+from squid_py.account import Account
+from squid_py.aquariuswrapper import AquariusWrapper
 from squid_py.config import Config
 from squid_py.keeper import Keeper
 from squid_py.log import setup_logging
-from squid_py.aquariuswrapper import AquariusWrapper
-from squid_py.account import Account
-
-from squid_py.utils import Web3Helper
 
 CONFIG_FILE_ENVIRONMENT_NAME = 'CONFIG_FILE'
 
 setup_logging()
+
 
 class Ocean:
     def __init__(self, config_file):
@@ -35,14 +33,14 @@ class Ocean:
         # For development, we use the HTTPProvider Web3 interface
         self._web3 = Web3(HTTPProvider(self.config.keeper_url))
 
-         # With the interface loaded, the Keeper node is connected with all contracts
+        # With the interface loaded, the Keeper node is connected with all contracts
         self.keeper = Keeper(self._web3, self.config.keeper_path, self.config.address_list)
 
         # Add the Metadata store to the interface
         if self.config.aquarius_url:
             self.metadata = AquariusWrapper(self.config.aquarius_url)
-        else: self.metadata = None
-
+        else:
+            self.metadata = None
 
         # Collect the accounts
         self.accounts = self.get_accounts()
@@ -50,7 +48,7 @@ class Ocean:
         assert self.accounts
 
     def print_config(self):
-        #TODO: Cleanup
+        # TODO: Cleanup
         print("Ocean object configuration:".format())
         print("Ocean.config.keeper_path: {}".format(self.config.keeper_path))
         print("Ocean.config.keeper_url: {}".format(self.config.keeper_url))
@@ -127,9 +125,8 @@ class Ocean:
         # 3) Publish to metadata store
         # Check if it's already registered first!
         if asset.asset_id in self.metadata.list_assets()['ids']:
-            #TODO: raise proper error
-            raise
-
+            # TODO: raise proper error
+            pass
         self.metadata.publish_asset_metadata(asset)
 
         # 4) Register the asset onto blockchain
@@ -139,4 +136,3 @@ class Ocean:
 class Order:
     def __init__(self):
         pass
-
