+++ squid_py/utils/__init__.py
@@ -1,11 +1,3 @@
-
-
-from .web3_helper_module import (
-    convert_to_bytes,
-    convert_to_string,
-    convert_to_text,
-)
-
-from .web3_helper import (
-    Web3Helper,
+from .utilities import (
+    Utilities
 )
